test(ui): move the quick bar layout test beside the code it covers
